backend/ai/upi_analyzer.py
import os
import re
import json
import io
from dotenv import load_dotenv
from google import genai
import pdfplumber

load_dotenv()

client = genai.Client(api_key=os.getenv("GEMINI_API_KEY"))


def extract_text_from_pdf(file_bytes: bytes) -> str:
    text = ""
    try:
        with pdfplumber.open(io.BytesIO(file_bytes)) as pdf:
            for page in pdf.pages:
                extracted = page.extract_text()
                if extracted:
                    text += extracted + "\n"
    except Exception:
        text = ""
    return text


def analyze_upi_statement(file_bytes: bytes, user_type: str = "gig") -> dict:
    raw_text = extract_text_from_pdf(file_bytes)

    if not raw_text or len(raw_text) < 100:
        return {
            "parsed": False,
            "error": "Could not extract text from the uploaded PDF. Please ensure it is a text-based UPI statement.",
        }

    truncated = raw_text[:3000]

    if user_type == "msme":
        focus = "business transaction patterns, revenue inflows, supplier payments, and cash flow consistency"
    else:
        focus = "income deposits, platform payouts (Swiggy, Zomato, Ola, Uber, Upwork etc.), spending patterns, and payment regularity"

    prompt = f"""
You are a financial analyst reviewing a 30-day UPI bank statement for a {user_type} in India.

Here is the raw statement text:
---
{truncated}
---

Analyze this statement and extract the following. Focus on {focus}.

Respond ONLY in this exact JSON format with no markdown, no extra text:
{{
  "total_credits": <total money received in INR as a number>,
  "total_debits": <total money spent in INR as a number>,
  "transaction_count": <total number of transactions as an integer>,
  "avg_credit": <average credit amount in INR as a number>,
  "income_sources": [<list of identified income/payment sources as strings>],
  "spending_categories": [<list of spending categories as strings>],
  "consistency_score": <a number 0-100 rating how consistent and regular the transactions are>,
  "risk_flags": [<list of any financial risk signals observed, empty list if none>],
  "summary": "<2 sentence plain English summary of this person's financial behavior from the statement>"
}}

If you cannot determine a value with confidence, use 0 for numbers and empty lists for arrays.
"""

    try:
        response = client.models.generate_content(
            model="gemini-2.0-flash",
            contents=prompt,
        )
        text = response.text.strip()
        if text.startswith("```"):
            text = text.split("```")[1]
            if text.startswith("json"):
                text = text[4:]
        parsed = json.loads(text.strip())
        parsed["parsed"] = True
        return parsed
    except Exception:
        amounts = re.findall(r"(?:₹|Rs\.?)\s?([\d,]+\.?\d*)", raw_text)
        amounts_clean = []
        for a in amounts:
            try:
                amounts_clean.append(float(a.replace(",", "")))
            except ValueError:
                continue

        total = sum(amounts_clean)
        count = len(amounts_clean)
        avg = round(total / count, 2) if count > 0 else 0
        consistency = min(100, int((count / 30) * 100))

        return {
            "parsed": True,
            "total_credits": round(total * 0.6, 2),
            "total_debits": round(total * 0.4, 2),
            "transaction_count": count,
            "avg_credit": avg,
            "income_sources": ["UPI Transfers"],
            "spending_categories": ["General Expenses"],
            "consistency_score": consistency,
            "risk_flags": [],
            "summary": f"Statement shows {count} transactions totaling approximately ₹{round(total):,} over 30 days.",
        }