backend/requirements.txt
fastapi
uvicorn
pdfplumber
pandas
python-multipart
pydantic
numpy
google-generativeai
python-dotenv
 